bot/requirements.txt
flask==3.0.3
schedule==1.2.1
beautifulsoup4==4.12.3
playwright==1.47.0
python-dotenv==1.0.1

